# openapi.yaml
openapi: 3.1.0
info:
  title: Sleeper Backend API
  description: |
    A comprehensive fantasy football API that aggregates data from **KeepTradeCut (KTC)** and **Sleeper API**.

    This API provides comprehensive fantasy football data by combining:
    - 🏈 **KeepTradeCut (KTC)**: Player rankings and trade values (Dynasty & Redraft rankings, 1QB & Superflex league formats, TEP scoring variations)
    - 🏈 **Sleeper API**: Player data & league management (Comprehensive player profiles, League management, Research data and projections)

    ## Performance Optimizations
    - **Database Caching Strategy**: First API call fetches from external APIs and caches in database, subsequent calls served instantly from local cache
    - **Response Time Guidelines**: First refresh call takes 30-60 seconds, cached data retrieval < 1 second, health checks < 100ms
    - **Gunicorn Multi-Worker Setup**: 9 worker processes for concurrent request handling with load balancing
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/rawarne2/sleeper-backend
servers:
  - url: http://localhost:5001
    description: Development server
  - url: https://sleeper-backend.vercel.app
    description: Production server
...
paths:
...
  /api/sleeper/league/{league_id}:
    get:
      tags:
        - Sleeper Leagues
      summary: Get comprehensive league data
      description: |
        Get comprehensive league data (league info, rosters, users). Returns complete league configuration, all rosters with player IDs, and user information. Checks database first, falls back to Sleeper API.
      operationId: getLeagueData
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League data retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueDataResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Sleeper Leagues
      summary: Refresh league data
      description: Refresh all league data from Sleeper API (league info, rosters, users)
      operationId: refreshLeagueData
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League data refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueRefreshResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Sleeper Leagues
      summary: Update league data
      description: Refresh all league data from Sleeper API (league info, rosters, users)
      operationId: updateLeagueData
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: League data refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueRefreshResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sleeper/league/{league_id}/rosters:
    get:
      tags:
        - Sleeper Leagues
      summary: Get league rosters
      description: Get league rosters data for all teams in the league
      operationId: getLeagueRosters
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Rosters retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RostersResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sleeper/league/{league_id}/users:
    get:
      tags:
        - Sleeper Leagues
      summary: Get league users
      description: Get league users data for all participants in the league
      operationId: getLeagueUsers
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/dashboard/league/{league_id}:
    get:
      tags:
        - Dashboard
      summary: Get dashboard league bundle
      description: |
        Returns a merged dashboard payload for a league, including rosters, player projections/values,
        research metadata, and optional per-player season stats.
      operationId: getDashboardLeagueBundle
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/SeasonQuery"
        - $ref: "#/components/parameters/IsRedraft"
        - $ref: "#/components/parameters/LeagueFormat"
        - $ref: "#/components/parameters/TepLevel"
      responses:
        "200":
          description: Dashboard payload retrieved successfully
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    IsRedraft:
      name: is_redraft
      in: query
      description: Ranking type - 'true' for redraft (current season only), 'false'
        for dynasty (long-term player value)
      required: false
      schema:
        type: string
        enum: [ "true", "false" ]
        default: "false"
      example: "false"
    LeagueFormat:
      name: league_format
      in: query
      description: League format - '1qb' for standard leagues, 'superflex' for
        superflex leagues (can start 2 QBs)
      required: false
      schema:
        type: string
        enum: [ "1qb", "superflex" ]
        default: "1qb"
      example: "superflex"
    TepLevel:
      name: tep_level
      in: query
      description: |
        TEP (Tight End Premium) scoring level:
        - '' (empty): Standard scoring
        - 'tep': +0.5 points per TE reception
        - 'tepp': +1.0 points per TE reception
        - 'teppp': +1.5 points per TE reception
      required: false
      schema:
        type: string
        enum: [ "", "tep", "tepp", "teppp" ]
        default: ""
      example: "tep"
    LeagueId:
      name: league_id
      in: path
      description: The Sleeper league ID
      required: true
      schema:
        type: string
        pattern: "^[0-9]+$"
      example: "1210364682523656192"
...
    SeasonQuery:
      name: season
      in: query
      description: The NFL season year
      required: false
      schema:
        type: string
        pattern: "^[0-9]{4}$"
        default: "2024"
      example: "2024"
...
  schemas:
...
    ErrorResponse:
      type: object
      description: >
        Standard JSON error envelope. Optional fields depend on the route;
        clients should tolerate unknown keys.
      required:
        - status
        - error
        - timestamp
      additionalProperties: true
      properties:
        status:
          type: string
          enum: [ "error" ]
        error:
          type: string
          description: Error message
        details:
          type: string
          description: >
            Additional detail as text. When the source value is structured
            (e.g. upstream API error object), the server may JSON-encode it into this string.
        timestamp:
          type: string
          format: date-time
          description: Error timestamp (RFC 3339, UTC with Z)
        hint:
          type: string
          description: Operator hint (e.g. maintenance auth)
        suggestion:
          type: string
          description: Suggested remediation (e.g. which refresh to run)
        job_id:
          type: string
        league_id:
          type: string
        season:
          type: string
        week:
          oneOf:
            - type: integer
            - type: string
          description: Week number or the literal "all" for research refresh errors
        league_type:
          type: string
          enum: [ "dynasty", "redraft" ]
        parameters:
          type: object
          additionalProperties: true
          description: Request parameters associated with the error (e.g. KTC format flags)
        configuration:
          type: object
          additionalProperties: true
          description: Server-side configuration context (e.g. cleanup target)
        results:
          type: array
          description: Per-item outcomes (e.g. prewarm summary entries)
          items:
            type: object
            additionalProperties: true
        sleeper_players_fetched:
          type: integer
          description: Count from Sleeper merge failures
...
    League:
      type: object
      properties:
        id:
          type: integer
        league_id:
          type: string
          description: Sleeper league ID
        name:
          type: string
          description: League name
        season:
          type: string
          description: Season year
        roster_positions:
          type: array
          items:
            type: string
          description: Required roster positions
        scoring_settings:
          type: object
          description: League scoring settings
        league_settings:
          type: object
          description: League configuration settings
        status:
          type: string
          description: League status
        draft_id:
          type: string
          description: Associated draft ID
        last_updated:
          type: string
          format: date-time
    Roster:
      type: object
      properties:
        id:
          type: integer
        league_id:
          type: string
        roster_id:
          type: integer
          description: Team/roster ID within league
        owner_id:
          type: string
          description: User ID of team owner
        players:
          type: array
          items:
            type: string
          description: Array of player IDs on roster
        starters:
          type: array
          items:
            type: string
          description: Array of starting player IDs
        reserve:
          type: array
          items:
            type: string
          description: Array of reserve player IDs
        taxi:
          type: array
          items:
            type: string
          description: Array of taxi squad player IDs
        settings:
          type: object
          description: Roster-specific settings
        last_updated:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: integer
        league_id:
          type: string
        user_id:
          type: string
          description: Sleeper user ID
        username:
          type: string
          description: Sleeper username
        display_name:
          type: string
          description: Display name
        avatar:
          type: string
          description: Avatar URL/ID
        team_name:
          type: string
          description: Fantasy team name
        user_metadata:
          type: object
          description: Additional user data
        last_updated:
          type: string
          format: date-time
    LeagueDataResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ "success" ]
        data:
          type: object
          properties:
            league:
              $ref: "#/components/schemas/League"
            rosters:
              type: array
              items:
                $ref: "#/components/schemas/Roster"
            users:
              type: array
              items:
                $ref: "#/components/schemas/User"
        source:
          type: string
          enum: [ "database", "sleeper_api" ]
        database_saved:
          type: boolean
        timestamp:
          type: string
          format: date-time
    RostersResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ "success" ]
        league_id:
          type: string
        rosters:
          type: array
          items:
            $ref: "#/components/schemas/Roster"
        count:
          type: integer
        last_updated:
          type: string
          format: date-time
        timestamp:
          type: string
          format: date-time
    UsersResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ "success" ]
        league_id:
          type: string
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        count:
          type: integer
        last_updated:
          type: string
          format: date-time
        timestamp:
          type: string
          format: date-time
    LeagueRefreshResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ "success", "partial_success", "error" ]
        timestamp:
          type: string
          format: date-time
        league_id:
          type: string
        league_data:
          type: object
          properties:
            status:
              type: string
            save_result:
              type: object
        users_data:
          type: object
          properties:
            status:
              type: string
            message:
              type: string
        rosters_data:
          type: object
          properties:
            status:
              type: string
            message:
              type: string
        errors:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              error:
                type: string
...